requirements.txt
pypdf # Loading PDFs
pytest # Unit Testing
fastapi
uvicorn
debugpy
langchain-community #LLM Library
langchain #LLM Library
langchainhub
gpt4all
chromadb # Vector Storage
requests
beautifulsoup4aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
backoff
bcrypt
beautifulsoup4
build
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
dataclasses-json
debugpy
Deprecated
dnspython
email_validator
fastapi
fastapi-cli
filelock
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
gpt4all
greenlet
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
iniconfig
Jinja2
jsonpatch
jsonpointer
kubernetes
langchain
langchain-community
langchain-core
langchain-text-splitters
langchainhub
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
numpy
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pluggy
posthog
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pypdf
PyPika
pyproject_hooks
pytest
python-dateutil
python-dotenv
python-multipart
PyYAML
requests
requests-oauthlib
rich
rsa
setuptools
shellingham
six
sniffio
soupsieve
SQLAlchemy
starlette
sympy
tenacity
tokenizers
tqdm
typer
types-requests
typing-inspect
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
